Removing myself from `CODEOWNERS` (#109)
I don't have enough expertise to review these PRs and the way this file is configured I get added to review all PRs to the repo.

Instead, if there's a PR that I should review, please tag me. This way, I'll know to only react to the ones that are relevant.